+++ example/pages/datetimePickerView/index.js
@@ -32,6 +32,6 @@ Page({
     }
   },
   onChange ({ detail }) {
-    Toast(`当前选中时间: ${detail}`)
+    Toast(`当前选中时间: ${detail.value}`)
   }
 })